+++ app/Models/User.php
@@ -13,8 +13,8 @@ use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
 use Spatie\Permission\Traits\HasRoles;
 
-#[Fillable(['name', 'email', 'password', 'status', 'bio', 'phone', 'country', 'city_state', 'postal_code', 'tax_id', 'facebook', 'x_url', 'linkedin', 'instagram', 'locale', 'avatar'])]
-#[Hidden(['password', 'remember_token'])]
+#[Fillable(['name', 'email', 'password', 'status', 'bio', 'phone', 'country', 'city_state', 'postal_code', 'tax_id', 'facebook', 'x_url', 'linkedin', 'instagram', 'locale', 'avatar', 'two_factor_secret', 'two_factor_confirmed_at'])]
+#[Hidden(['password', 'remember_token', 'two_factor_secret'])]
 class User extends Authenticatable implements MustVerifyEmail
 {
     /** @use HasFactory<UserFactory> */
@@ -23,8 +23,9 @@ class User extends Authenticatable implements MustVerifyEmail
     protected function casts(): array
     {
         return [
-            'email_verified_at' => 'datetime',
-            'password'          => 'hashed',
+            'email_verified_at'       => 'datetime',
+            'two_factor_confirmed_at' => 'datetime',
+            'password'                => 'hashed',
         ];
     }
 }